Return the image-to-profile decision to imageio/
Six functions, ~630 lines, out of common/colorspaces.c into a new
src/imageio/imageio_profile.{c,h}: dt_image_find_best_color_profile,
dt_colorspaces_get_input_profile_from_image, dt_colorspaces_get_embedded_profile,
dt_colorspaces_get_output_profile, dt_colorspaces_cicp_to_type and the helper behind them.

They look like colour management and are not. Deciding that a JPEG's embedded profile beats
its EXIF colorspace tag, that an AVIF's CICP block maps to Rec.2020, or that a RAW should use
its camera matrix, is knowledge of what each FORMAT stores and where. That is codec work, and
it was the only reason a layer-1 colour module included six imageio codec headers -- four of
them conditionally, on HAVE_OPENJPEG / HAVE_LIBAVIF / HAVE_LIBHEIF.

What stays: everything that builds, describes or applies a profile. colorspaces.c keeps the
LittleCMS2 work and the application-wide profile list, and no longer knows a codec exists.

One new function, because the moved code needed something the file-static _create_profile
could no longer give it: dt_colorspaces_new_image_profile(type, profile, owns_profile). It is
the constructor half of the pair whose destructor landed two commits ago, and it puts the
ownership question in the signature -- which is exactly the question that made the previous
commit abort every CI runner.

Layering violations 215 -> 209, the predicted figure.

Also, one thing that surfaced while checking po/POTFILES.in for this move and turned out to be
worse than a bookkeeping detail: 24 of its entries pointed at files that no longer exist, and
nearly all of them are files THIS refactor series moved -- bauhaus.c, thumbtable.c,
accelerators.c, printprof.c, iop_order.c, darktable.c and the rest. Nothing validates that
file, so each move silently dropped its strings from ansel.pot. Repathing them returns ~355
translatable strings to translators. Also removed one duplicate entry, which predates this.

Four configurations build clean, all four static gates pass, and tools/check_it_runs.sh exports
cleanly -- the check that would have caught the last commit's double free.
