+++ po/POTFILES.in
@@ -87,38 +87,36 @@ data/photos.ansel.ansel.appdata.xml.in
 data/photos.ansel.ansel.desktop.in
 data/anselconfig.dtd
 data/anselconfig.xml.in
-src/bauhaus/bauhaus.c
-src/chart/main.c
-src/cli/main.c
-src/common/chromatic_adaptation.h
+src/widgets/bauhaus.c
+src/apps/ansel-chart/main.c
+src/apps/ansel-cli/main.c
+src/pixel/chromatic_adaptation.h
 src/common/collection.c
 src/common/color_vocabulary.c
 src/common/colorlabels.c
 src/common/colorspaces.c
-src/common/color_vocabulary.c
 src/common/colorchecker.h
 src/common/colorchecker.c
 src/common/cups_print.c
-src/common/darktable.c
-src/common/darktable.h
+src/darktable.c
+src/darktable.h
 src/common/database.c
 src/common/debug.h
-src/common/eigf.h
+src/pixel/eigf.h
 src/common/exif.cc
-src/common/fast_guided_filter.h
+src/pixel/fast_guided_filter.h
 src/common/film.c
 src/common/history_actions.c
-src/common/history_merge_gui.c
+src/gui/develop/history_merge_gui.c
 src/common/history.c
 src/common/http_server.c
 src/common/image_cache.c
 src/common/image.c
 src/common/image.h
 src/common/imagebuf.c
-src/common/imageio.c
-src/common/import.c
-src/common/iop_order.c
-src/common/iop_profile.c
+src/imageio/imageio_core.c
+src/gui/import.c
+src/develop/iop_order.c
 src/common/map_locations.c
 src/common/metadata.c
 src/common/mipmap_cache.c
@@ -127,7 +125,7 @@ src/common/opencl.c
 src/common/pdf.h
 src/common/points.h
 src/common/printing.c
-src/common/printprof.c
+src/colorprofiles/printprof.c
 src/common/pwstorage/pwstorage.c
 src/common/ratings.c
 src/common/sentry.c
@@ -159,15 +157,15 @@ src/develop/masks/masks.c
 src/develop/masks/polygon.c
 src/develop/pixelpipe_hb.c
 src/develop/tiling.c
-src/dtgtk/resetlabel.c
-src/dtgtk/preview_window.c
-src/dtgtk/thumbnail.c
-src/dtgtk/thumbnail_btn.c
-src/dtgtk/thumbtable.c
-src/generate-cache/main.c
-src/gui/accelerators.c
-src/gui/gtk.c
-src/gui/gtkentry.c
+src/widgets/resetlabel.c
+src/gui/dtgtk/preview_window.c
+src/gui/dtgtk/thumbnail.c
+src/widgets/thumbnail_btn.c
+src/gui/dtgtk/thumbtable.c
+src/apps/ansel-generate-cache/main.c
+src/widgets/accelerators.c
+src/gui/application.c
+src/widgets/gtkentry.c
 src/gui/guides.c
 src/gui/hist_dialog.c
 src/gui/preferences.c
@@ -219,7 +217,7 @@ src/iop/cacorrectrgb.c
 src/iop/censorize.c
 src/iop/channelmixer.c
 src/iop/channelmixerrgb.c
-src/iop/choleski.h
+src/math/choleski.h
 src/iop/clipping.c
 src/iop/colisa.c
 src/iop/colorbalance.c
@@ -324,7 +322,7 @@ tools/release/create_release.sh
 tools/release/devel-stats.sh
 tools/release/generate-authors.sh
 tools/generate_prefs.xsl
-src/common/lut_viewer.c
+src/gui/lut_viewer.c
 src/develop/dev_pixelpipe.c
 src/develop/pixelpipe_raster_masks.c
 src/iop/basebuffer.c
