+++ src/iop/colorout.c
@@ -81,6 +81,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include <string.h>
+#include "imageio/imageio_profile.h"
 
 // max iccprofile file name length
 // must be in synch with dt_colorspaces_color_profile_t
